Make an ant's life less intense
CPU was high even without rendering.

Each ant didn't need to take 100 steps a second.
It should be enough to take a step once a second
and to eat each 5 seconds.
Also, each time it acts, it has a 0,2% chance to die.
Interestingly, there's still a ton of CPU usage. Maybe Thread.sleep is
not the way? Gotta use the async profiler.